fix -X and -I options
Previously when executing "filterfs /src/dir /dest/dir -X '*.bak'" this
error was generated:

error: general error while parsing options.

but "filterfs /src/dir /dest/dir --exclude='*.bak'" worked as expected.

The cause was that when processing the '-X' option, it was expected the
string to be passed to str_consume was of the form '-X %s', but the
whitespace between '-X' and '%s' was eaten up previously when parsing
parameters, so the resulting strings would be '-X' and '%s'.

We remove the wrong assumption and the options now work as expected.
Both '-X%s' and '-X %s' work now as expected; %s is of the form
'pattern[:pattern[...]]'. The same applies for the '-I' option.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>